feat: rename poor_mans_cms into metas, will maybe rollback at some point
